@@ -7,41 +7,44 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emotion/jest": "^11.2.1",
-    "@emotion/react": "^11.4.1",
-    "@emotion/styled": "^11.3.0",
-    "@orfium/ictinus": "^3.14.1",
-    "@testing-library/jest-dom": "^5.11.9",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.1.3",
-    "@types/jest": "^26.0.23",
-    "@types/node": "^14.14.28",
-    "@types/react": "^17.0.3",
-    "@types/react-dom": "^17.0.0",
-    "@types/react-final-form-listeners": "^1.0.0",
-    "@types/react-router-dom": "^5.1.7",
-    "axios": "^0.21.1",
+    "@emotion/jest": "^11.13.0",
+    "@emotion/react": "11.14.0",
+    "@emotion/styled": "11.14.1",
+    "@fontsource/roboto": "5.2.6",
+    "@mui/icons-material": "7.3.1",
+    "@mui/material": "7.3.1",
+    "@mui/styles": "6.5.0",
+    "@testing-library/jest-dom": "^6.6.4",
+    "@testing-library/react": "^16.3.0",
+    "@testing-library/user-event": "^14.6.1",
+    "@types/jest": "^30.0.0",
+    "@types/node": "^24.2.0",
+    "@types/react": "^19.1.9",
+    "@types/react-dom": "^19.1.7",
+    "@types/react-final-form-listeners": "^1.0.3",
+    "axios": "^1.11.0",
     "customize-cra": "^1.0.0",
-    "express": "^4.21.2",
-    "final-form": "^4.20.2",
-    "final-form-arrays": "^3.0.2",
-    "history": "^4.10.1",
+    "express": "^5.1.0",
+    "final-form": "^5.0.0",
+    "final-form-arrays": "^4.0.0",
+    "history": "^5.3.0",
     "lodash": "^4.17.21",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-final-form": "^6.5.3",
-    "react-final-form-arrays": "^3.1.3",
-    "react-final-form-listeners": "^1.0.3",
-    "react-query": "^3.16.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "typescript": "^4.1.3",
-    "universal-cookie": "^4.0.4",
-    "webpack": "4.44.2"
+    "npm-check-updates": "18.0.2",
+    "polished": "4.3.1",
+    "react": "^19.1.1",
+    "react-dom": "^19.1.1",
+    "react-final-form": "^7.0.0",
+    "react-final-form-arrays": "^4.0.0",
+    "react-final-form-listeners": "^3.0.0",
+    "react-query": "^3.39.3",
+    "react-router-dom": "^7.7.1",
+    "react-scripts": "5.0.1",
+    "typescript": "^5.9.2",
+    "universal-cookie": "^8.0.1",
+    "webpack": "5.101.0"
   },
   "scripts": {
     "start:prod": "node --max_old_space_size=460 server.js",
-    "start": "react-app-rewired start",
     "start": "react-scripts --openssl-legacy-provider start",
     "start_win": "react-scripts start",
     "build": "react-app-rewired --openssl-legacy-provider build",
