@@ -5,39 +5,59 @@ import Login from 'pages/Login';
 import PatientDirectory from 'pages/PatientDirectory';
 import RegisterPatient from 'pages/RegisterPatient';
 import Settings from 'pages/Settings';
-import { Redirect, Switch } from 'react-router-dom';
+import { Navigate, createBrowserRouter, RouterProvider } from 'react-router-dom';
 
-import PrivateRoute from './PrivateRoute';
-import PublicRoute from './PublicRoute';
 import urls from './urls';
 import EpisodeDetails from '../pages/EpisodeDetails';
 import PatientDetails from '../pages/PatientDetails';
 import RegisterEpisode from '../pages/RegisterEpisode';
+import { __TOKEN__ } from '../utils/constants';
+import { getUserStorageItem } from '../utils/storage';
 
-const Routes: React.FC = () => (
-  <Switch>
-    <PublicRoute exact path={urls.login()} component={Login} />
-    <PrivateRoute exact path={urls.settings()} component={Settings} />
-    <PrivateRoute exact path={urls.registerPatient()} component={RegisterPatient} />
-    <PrivateRoute
-      exact
-      path={`${urls.patients()}/:hospitalID/:patientID${urls.episodes()}/:episodeID`}
-      component={EpisodeDetails}
-    />
-    <PrivateRoute
-      exact
-      path={`${urls.patients()}/:hospitalID/:patientID/add-episode`}
-      component={RegisterEpisode}
-    />
-    <PrivateRoute
-      exact
-      path={`${urls.patients()}/:hospitalID/:patientID`}
-      component={PatientDetails}
-    />
-    <PrivateRoute exact path={[urls.patients()]} component={PatientDirectory} />\
-    <PrivateRoute exact path={urls.landingPage()} component={LandingPage} />
-    <Redirect to={urls.login()} />
-  </Switch>
-);
+const PrivateRoute = ({ children }: { children: React.ReactNode }) => {
+  const token = getUserStorageItem(__TOKEN__);
+  return token ? <>{children}</> : <Navigate to={urls.login()} />;
+};
+
+const router = createBrowserRouter([
+  {
+    path: urls.login(),
+    element: <Login />,
+  },
+  {
+    path: urls.settings(),
+    element: <PrivateRoute><Settings /></PrivateRoute>,
+  },
+  {
+    path: urls.registerPatient(),
+    element: <PrivateRoute><RegisterPatient /></PrivateRoute>,
+  },
+  {
+    path: `${urls.patients()}/:hospitalID/:patientID${urls.episodes()}/:episodeID`,
+    element: <PrivateRoute><EpisodeDetails /></PrivateRoute>,
+  },
+  {
+    path: `${urls.patients()}/:hospitalID/:patientID/add-episode`,
+    element: <PrivateRoute><RegisterEpisode /></PrivateRoute>,
+  },
+  {
+    path: `${urls.patients()}/:hospitalID/:patientID`,
+    element: <PrivateRoute><PatientDetails /></PrivateRoute>,
+  },
+  {
+    path: urls.patients(),
+    element: <PrivateRoute><PatientDirectory /></PrivateRoute>,
+  },
+  {
+    path: urls.landingPage(),
+    element: <PrivateRoute><LandingPage /></PrivateRoute>,
+  },
+  {
+    path: '*',
+    element: <Navigate to={urls.login()} />,
+  },
+]);
+
+const Routes: React.FC = () => <RouterProvider router={router} />;
 
 export default Routes;
