@@ -2,11 +2,5 @@ import styled from '@emotion/styled';
 
 export const Wrapper = styled.div`
   display: flex;
-  margin-right: 10px;
-
-  input {
-    height: 20px;
-    margin-right: 10px;
-    width: 20px;
-  }
+  gap: 8px;
 `;
