@@ -1,27 +1,31 @@
 import { useEffect, useState } from 'react';
 
-import useBreakpoints from '@orfium/ictinus/dist/hooks/useBreakpoints';
+import { useTheme } from '@mui/material/styles';
+import useMediaQuery from '@mui/material/useMediaQuery';
 
 export const useResponsiveLayout = () => {
   const [expanded, setExpanded] = useState(false);
-  const breakpoints = useBreakpoints();
+  const theme = useTheme();
 
-  const isDesktop =
-    breakpoints.des1200 || breakpoints.des1440 || breakpoints.des1366 || breakpoints.des1920;
+  const isDesktop = useMediaQuery(theme.breakpoints.up('des1200'));
+  const isSmallDesktop = useMediaQuery(theme.breakpoints.between('des1200', 'des1366'));
+  const isMediumDesktop = useMediaQuery(theme.breakpoints.between('des1366', 'des1440'));
+  const isLargeDesktop = useMediaQuery(theme.breakpoints.between('des1440', 'des1920'));
+  const isXLargeDesktop = useMediaQuery(theme.breakpoints.up('des1920'));
 
   useEffect(() => {
-    setExpanded(breakpoints.des1440 && !breakpoints.des1920);
-  }, [breakpoints]);
+    setExpanded(isLargeDesktop);
+  }, [isLargeDesktop]);
 
   const toggle = () => {
     setExpanded((prev) => !prev);
   };
 
   const responsiveProps = {
-    isSmallDesktop: breakpoints.des1200 && !breakpoints.des1366,
-    isMediumDesktop: breakpoints.des1366 && !breakpoints.des1440,
-    isLargeDesktop: breakpoints.des1440 && !breakpoints.des1920,
-    isXLargeDesktop: breakpoints.des1920,
+    isSmallDesktop,
+    isMediumDesktop,
+    isLargeDesktop,
+    isXLargeDesktop,
   };
 
   return {
