@@ -1,5 +1,3 @@
-import { NotificationTypes } from '@orfium/ictinus/dist/components/Notification/Notification';
-
 import { NotificationType } from '../components/Notifications/Notifications';
 import { addNotification } from '../providers/Notifications/actions';
 import { useNotifications } from '../providers/Notifications/NotificationProvider';
@@ -7,7 +5,7 @@ import { useNotifications } from '../providers/Notifications/NotificationProvide
 export const useSetNotification = () => {
   const [, dispatch] = useNotifications();
 
-  return (message: string, type: NotificationTypes, isGlobal = false, isPreview = false) => {
+  return (message: string, type: NotificationType['type'], isGlobal = false, isPreview = false) => {
     const newNotification: NotificationType = {
       isGlobal,
       isPreview,
