@@ -1,14 +1,12 @@
 import styled from '@emotion/styled';
 
-import { flex } from '../../theme/functions';
-
 export const Container = styled.div`
   overflow-y: auto;
-  padding: ${({ theme }) => theme.spacing.md};
+  padding: ${({ theme }) => theme.spacing(4)};
 `;
 
 export const PageWrapper = styled.div<{ isDesktop: boolean }>`
-  ${flex};
+  display: flex;
   flex-direction: column;
   height: 100%;
   width: ${({ isDesktop }) => (isDesktop ? '50%' : '100%')};
