@@ -1,10 +1,8 @@
 import styled from '@emotion/styled';
 
-import { flex } from '../../../../../theme/functions';
-
 export const FieldContainer = styled.div`
+  display: flex;
   flex-direction: column;
-  ${flex};
   gap: 16px;
   margin-bottom: 24px;
   margin-top: 8px;
