@@ -1,36 +1,30 @@
 import styled from '@emotion/styled';
-import { transition } from '@orfium/ictinus/dist/theme/functions';
 import { rem } from 'polished';
 
 export const ListItem = styled.li<{ isOpen: boolean }>`
   background: ${({ theme, isOpen }) =>
-    isOpen ? theme.utils.getColor('lightCoolGray', 400) : '#fff'};
-  border-radius: ${({ theme }) => theme.spacing.xsm};
-  border-radius: ${({ theme }) => theme.spacing.xsm};
+    isOpen ? theme.palette.grey[300] : '#fff'};
+  border-radius: ${({ theme }) => theme.spacing(2)};
   box-sizing: border-box;
   cursor: pointer;
   list-style: none;
   outline: ${rem(1)} solid
-    ${({ theme, isOpen }) =>
-      theme.utils.getColor(isOpen ? 'blue' : 'lightGray', isOpen ? 300 : 400)};
-  :not(:last-of-type) {
-    margin-bottom: ${({ theme }) => `${theme.spacing.md}`};
-  }
+    ${({ theme, isOpen }) => isOpen ? theme.palette.primary.dark : theme.palette.grey[300]};
+  padding: ${({ theme }) => `${theme.spacing(4)}`};
+  transition: all 0.1s ease-in-out;
 
   :hover {
-    outline: ${rem(2)} solid ${({ theme }) => theme.utils.getColor('lightGray', 300)};
+    outline: ${rem(2)} solid ${({ theme }) => theme.palette.grey[300]};
   }
-  ${transition(0.1)};
-  padding: ${({ theme }) => `${theme.spacing.md}`};
+  :not(:last-of-type) {
+    margin-bottom: ${({ theme }) => `${theme.spacing(4)}`};
+  }
 
   section[aria-expanded='true'] {
     background-color: transparent;
     max-height: 1000px;
     overflow: auto;
-    ${transition(0.2)};
-  }
-  section[aria-expanded='false'] {
-    ${transition(0.2)};
+    transition: max-height 0.2s ease-in-out;
   }
 `;
 
@@ -38,19 +32,16 @@ export const Header = styled.header<{
   isOpen: boolean;
 }>`
   align-items: center;
-  color: ${({ theme }) => theme.utils.getColor('lightGray', 600)};
+  color: ${({ theme }) => theme.palette.grey[400]};
   display: flex;
   justify-content: space-between;
-  > button > span > svg {
-    ${transition(0.2)};
-
-    transform: rotate(${({ isOpen }) => (isOpen ? `180deg` : 0)});
-  }
-
-  ${transition(0.2)};
-
   margin-bottom: ${({ isOpen }) => (isOpen ? '20px' : '0')};
   padding: ${({ isOpen }) => (isOpen ? '1px' : '0')};
+  transition: all 0.1s ease-in-out;
+  > button > span > svg {
+    transform: rotate(${({ isOpen }) => (isOpen ? `180deg` : 0)});
+    transition: all 0.1s ease-in-out;
+  }
 `;
 
 export const InternalContainer = styled.section<{
@@ -63,6 +54,6 @@ export const InternalContainer = styled.section<{
   ::-webkit-scrollbar {
     display: none;
   }
-  ${transition(0.2)};
   scrollbar-width: none;
+  transition: margin-bottom 0.2s ease-in-out;
 `;
