@@ -5,5 +5,5 @@ export const ComponentWrapper = styled.div`
   box-sizing: border-box;
   height: ${rem(411)};
   overflow-y: auto;
-  padding: ${({ theme }) => theme.spacing.md};
+  padding: ${({ theme }) => theme.spacing(4)};
 `;
