@@ -17,7 +17,7 @@ export const IconButtonWrapper = styled.div`
   right: 4%;
 
   button {
-    box-shadow: ${(props) => props.theme.elevation['02']};
+    box-shadow: '0 4px 6px -1px rgba(0,0,0,0.1), 0 2px 4px -1px rgba(0,0,0,0.06)';
   }
 
   svg,
