@@ -1,5 +1,4 @@
-import { FilterOption } from '@orfium/ictinus/dist/components/Filter/types';
-
+import { FilterOption } from '../types';
 import { RegisterPatientFormType } from './types';
 import { HospitalsAPI } from '../../models/apiTypes';
 
@@ -13,15 +12,7 @@ export const getHospitalOptions = (hospitals: HospitalsAPI[]): FilterOption[] =>
 };
 
 export const patientFormValidation = (values: RegisterPatientFormType) => {
-  const errors = {} || {
-    hospital: '',
-    firstName: '',
-    lastName: '',
-    yearOfBirth: '',
-    gender: '',
-    patientHospitalId: '',
-    phone1: '',
-  };
+  const errors: Partial<Record<keyof RegisterPatientFormType, string>> = {};
 
   if (!values.hospital && typeof values.hospital !== 'object') {
     errors.hospital = REQUIRED_FIELD_MSG;
