@@ -1,20 +1,22 @@
-import React, { FC } from 'react';
+import React, { FC, PropsWithChildren } from 'react';
 
-import { ThemeProvider as OrfiumThemeProvider } from '@orfium/ictinus';
+
+import { ThemeProvider } from '@mui/material/styles';
 import { render } from '@testing-library/react';
 import { BrowserRouter as Router } from 'react-router-dom';
 import cookies from 'utils/cookies';
 
+import theme from '../theme/theme';
 import { __TOKEN__ } from '../utils/constants';
 export const LOGIN_TOKEN_MOCK = 'test-token';
 
-export const MockProviders: FC = ({ children }) => (
+export const MockProviders: FC<PropsWithChildren> = ({ children }) => (
   <Router>
-    <OrfiumThemeProvider theme={{}}>{children}</OrfiumThemeProvider>
+    <ThemeProvider theme={theme}>{children}</ThemeProvider>
   </Router>
 );
 
-const renderWithProviders = (children: JSX.Element, fakeLogin = false) => {
+const renderWithProviders = (children: React.ReactNode, fakeLogin = false) => {
   // logout on render
   cookies.remove(__TOKEN__);
 
